Correcting method name '_keyPressed' in UITextField and UITextView inherited from UIResponder

+++ Frameworks/include/UIResponderInternal.h
@@ -17,8 +17,8 @@
 
 @interface UIResponder ()
 
-+ (void)_keyPressed:(unsigned)key;
-- (void)_keyPressed:(unsigned)key;
++ (void)_keyPressed:(unsigned short)key;
+- (void)_keyPressed:(unsigned short)key;
 + (id)_deleteRange:(id)num;
 
 @end
